+++ PlayniteVndbExtension/DescriptionFormatter.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using Playnite.SDK;
+
+namespace VndbSharp
+{
+    public class DescriptionFormatter
+    {
+        public string Format(string description)
+        {
+            var formatted = description.Replace("\n", "<br>" + System.Environment.NewLine);
+            return formatted;
+        }
+    }
+}